+++ src/main/java/plugily/projects/buildbattle/handlers/LanguageMigrator.java
@@ -19,17 +19,13 @@
 
 package plugily.projects.buildbattle.handlers;
 
+import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.FileConfiguration;
 import plugily.projects.buildbattle.Main;
-import plugily.projects.minigamesbox.classic.PluginMain;
 import plugily.projects.minigamesbox.classic.utils.configuration.ConfigUtils;
 import plugily.projects.minigamesbox.classic.utils.migrator.MigratorUtils;
 
 import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.logging.Level;
 
 /*
   NOTE FOR CONTRIBUTORS - Please do not touch this class if you don't know how it works! You can break migrator modifying these values!
@@ -38,7 +34,7 @@ import java.util.logging.Level;
 /**
  * @author Tigerpanzer_02
  * <p>
- * Created at 21.09.2021
+ * Created at 07.2025
  */
 @SuppressWarnings("deprecation")
 public class LanguageMigrator {
@@ -46,7 +42,8 @@ public class LanguageMigrator {
   public enum PluginFileVersion {
     /*ARENA_SELECTOR(0),*/ BUNGEE(1), CONFIG(1), LANGUAGE(2),
     /*LEADERBOARDS(0),*/ MYSQL(1), PERMISSIONS(1), POWERUPS(1),
-    /*SIGNS(0),*/ SPECIAL_ITEMS(1), SPECTATOR(1)/*, STATS(0)*/;
+    /*SIGNS(0),*/ SPECIAL_ITEMS(1), SPECTATOR(1)/*, STATS(0)*/,
+    MAIN_MENU(1);
 
     private final int version;
 
@@ -69,6 +66,9 @@ public class LanguageMigrator {
   private void updatePluginFiles() {
     for(PluginFileVersion pluginFileVersion : PluginFileVersion.values()) {
       String fileName = pluginFileVersion.name().toLowerCase();
+      if(fileName.equalsIgnoreCase(PluginFileVersion.MAIN_MENU.name())) {
+        fileName = "heads/mainmenu";
+      }
       int newVersion = pluginFileVersion.getVersion();
       File file = new File(plugin.getDataFolder() + "/" + fileName + ".yml");
       FileConfiguration configuration = ConfigUtils.getConfig(plugin, fileName, false);
@@ -79,15 +79,14 @@ public class LanguageMigrator {
       if(oldVersion == newVersion) {
         continue;
       }
-      plugin.getDebugger().debug(Level.WARNING, "[System notify] The " + fileName + "  file is outdated! Updating...");
+      Bukkit.getLogger().info("[System notify] The " + fileName + "  file is outdated! Updating...");
       for(int i = oldVersion; i < newVersion; i++) {
         executeUpdate(file, pluginFileVersion, i);
       }
 
       updatePluginFileVersion(file, configuration, oldVersion, newVersion);
-      plugin.getDebugger().debug(Level.WARNING, "[System notify] " + fileName + " updated, no comments were removed :)");
-      plugin.getDebugger().debug(Level.WARNING, "[System notify] You're using latest " + fileName + " file now! Nice!");
-
+      Bukkit.getLogger().info("[System notify] " + fileName + " updated, no comments were removed :)");
+      Bukkit.getLogger().info("[System notify] You're using latest " + fileName + " file now! Nice!");
     }
   }
 
@@ -98,6 +97,147 @@ public class LanguageMigrator {
           case 1:
             MigratorUtils.insertAfterLine(file, "      Heads:", "        Database:\n" +
                 "          Lore: \"Get Head %value%\"");
+            break;
+          default:
+            break;
+        }
+      case MAIN_MENU:
+        switch(version) {
+          case 0:
+            MigratorUtils.addNewLines(file, "# Use HeadDatabase in addition to the configured ones\n" +
+                "# You can remove categories from the catalog which shouldn't be loaded!\n" +
+                "# Catalog:\n" +
+                "#      - alphabet\n" +
+                "#      - animals\n" +
+                "#      - blocks\n" +
+                "#      - decoration\n" +
+                "#      - food-drinks\n" +
+                "#      - humanoid\n" +
+                "#      - humans\n" +
+                "#      - miscellaneous\n" +
+                "#      - monsters\n" +
+                "#      - plants\n" +
+                "database-alphabet:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvYzI1MDhlMmNhNjUwMGJjZTMwNTM5YzM4ODg0MmE1NjcyYjdiYzI5YTY4NzZkZDZhNTAyNTY3MmUyNTJkMjVkYSJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: alphabet\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-animals:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvOTQyYjllYzM0OTYwNjRiMDg1ZjdkMzBjZTkwYTBjOGY3NjM2YzhlZjUzMDNiMjBjMjVjYTEwYTk5N2JkNzQzMyJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: animals\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-blocks:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvYzViNDQ4ZmQ5NWM4NzRjOTVmZTc0ODQ0NDFhNDM5NGQ2NDZiNzJiYzgyYTUyNzQ4M2ZkYzcwY2E3OTg2ZmNhNSJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: blocks\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-decoration:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvMjNhZjk0YTUwZmFmNTBhZTAyMzZjNzExZWMxMzZiMjgwOGFjODJiYjE3MWQ0MmIzOGQxNjc2MGQyMjBmMjU4MiJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: decoration\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-food-drinks:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvZDUxMzdmYzBjZDUxMjAyNjcyMzgyYmZhZTAxZmViM2UxZTJiNzMxMDdlOThkMmM2YzhmNzE5ZjFkYTUzMGU2OCJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: food-drinks\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-humanoid:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvOTc5Yzc3ZjEyYzg2NzM5OGE0MTgxMDgxZmI1YmY3MGM1ZmYzMjcxNWM2ODk3NTBjNmU3MTdkMDY5ZTgxMzFhOSJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: humanoid\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-humans:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvNDQ5MmRmZTNjZThjYjYxYjkwODYwZWZmOTM4Y2I4M2UxOWMxNmU5Y2IyZDliZjJhZDc4OTZjOTBmNWIyZTFmMCJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: humans\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-miscellaneous:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvNDQ3OTE3MTkzZTlmYjdjYjk2OGZiYTAzYTJlMjg0YTA0NjEyMWIyZGMwMTU0MWNhY2RiYTUzYTRmOWJiZjcwOCJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: miscellaneous\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-monsters:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvZTgyYjQxZjY3YjMzNzE0NWM1ZWI4M2I5ZTAwMTU1NzQ4MDE2MGE1NWU5MWVmODYzZWIwYmYwNTU3Mzg0MmNlMSJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: monsters\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-plants:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvNDA4MzY2YjgxMWM0MzU5YWVhNjY4NjMwMGNmZDQ1MGQ5ZWFhNDYxNGMzYWYwOGExN2YxNTEzZjVkMmY0OGM0YSJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: plants\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"\n" +
+                "database-search:\n" +
+                "  displayname: '&6%value% heads'\n" +
+                "  texture: \"eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvNjViOTVkYTEyODE2NDJkYWE1ZDAyMmFkYmQzZTdjYjY5ZGMwOTQyYzgxY2Q2M2JlOWMzODU3ZDIyMmUxYzhkOSJ9fX0=\"\n" +
+                "  lore:\n" +
+                "    - \"Click to open\"\n" +
+                "    - \"%value% menu\"\n" +
+                "  enabled: true\n" +
+                "  database: true\n" +
+                "  config: search\n" +
+                "  permission: buildbattle.heads\n" +
+                "  menuname: \"%value% blocks\"");
             break;
           default:
             break;