+++ .github/CHANGELOG.md
@@ -11,7 +11,6 @@ deploy webhooks to print changelog successfully
 That's all, matcher will stop when detects next line started with `###` match
 
 ### 5.1.0 Release (10.07.2025)
-* TODO MAKE LANGUAGE:YML UPDATE BEFORE RELEASE!
 * Added Heads from HeadDatabase (selected on each server start 200 per category)
 * Added Heads Search function
 * Added paginated Heads menu to support more than one site of heads
@@ -37,6 +36,7 @@ That's all, matcher will stop when detects next line started with `###` match
 * Changed Allow Plot selection also if ArenaState is FULL_GAME
 * Changed (for API usage) renamed BBTheme to GuessTheme
 * Changed Hide guessing players from builder in gtb mode
+* Changed LanguageMigrator on plugin level to update file versions
 * Updated Minigamesbox to 1.4.2
 
 ### 5.0.15 Release (03.05.2025)